move col row input into indepent dirctories, remove antd.ButtonGroup
