+++ components/col/index.js
@@ -1,0 +1,2 @@
+import { Col } from '../layout';
+export default Col;
