+++ components/form/index.jsx
@@ -1,11 +1,7 @@
 import Form from './Form';
 import FormItem from './FormItem';
 import ValueMixin from './ValueMixin';
-import Input from './Input';
 
 Form.Item = FormItem;
 Form.ValueMixin = ValueMixin;
-export default {
-  Form,
-  Input
-};
+export default Form;
