+++ components/row/index.js
@@ -1,0 +1,2 @@
+import { Row } from '../layout';
+export default Row;
