+++ index.js
@@ -36,13 +36,12 @@ const antd = {
   Menu: require('./components/menu'),
   Timeline: require('./components/timeline'),
   Button: require('./components/button'),
-  ButtonGroup: require('./components/button').Group,
   Icon: require('./components/icon'),
-  Row: require('./components/layout').Row,
-  Col: require('./components/layout').Col,
+  Row: require('./components/row'),
+  Col: require('./components/col'),
   Spin: require('./components/spin'),
-  Form: require('./components/form').Form,
-  Input: require('./components/form').Input,
+  Form: require('./components/form'),
+  Input: require('./components/input'),
   Calendar: require('./components/calendar'),
   Timepicker: require('./components/timepicker'),
 };
